v0.0

+++ helpers/useFunctions.py
@@ -1,8 +1,7 @@
 import numpy as np
 
 def use_sigmoid(x):
-   return 1/(1+np.exp(-x))
-
+    return float(1 / (1 + np.exp(-x)))
 
 def use_tanh(x):
-   return np.tanh(x)
+    return float(np.tanh(x))
